ARM: SMP: basic IPI triggered completion support
We need a mechanism to let an inbound CPU signal that it is alive before
even getting into the kernel environment i.e. from early assembly code.
Using an IPI is the simplest way to achieve that.

This adds some basic infrastructure to register a struct completion
pointer to be "completed" when the dedicated IPI for this task is
received.

Signed-off-by: Nicolas Pitre <nico@linaro.org>
(cherry picked from commit 5135d875e1457ef946a055003d8f80713e862135)
Signed-off-by: Alex Shi <alex.shi@linaro.org>

Conflicts:
	arch/arm/kernel/smp.c