mptcp: remove fully overlapped skbs from ofo queue
When inserting in the ofo queue, it is necessary to always delete skbs
that are fully covered by adding the new segments.  If we do not do
this, we cannot maintain the invariant that the ofo q sequence numbers
are sorted. An unsorted ofo q will lead to performance issues and bugs
in for example tcp_collapse, which relies on this invariant.

Fixes: c195e8345db7 (Use tcp_try_coalesce in the meta-rcv and ofo-queue)
Signed-off-by: Tim Froidcoeur <tim.froidcoeur@tessares.net>
Signed-off-by: Christoph Paasch <cpaasch@apple.com>
(cherry picked from commit fda73d868366e7e6220b5d3d106207bbd0e462de)
Signed-off-by: Christoph Paasch <cpaasch@apple.com>